Add 'ScreenController' factory function
To generate the UI for the 'tic-tac-toe' game, and let players play the
game via the DOM rather than the console, the ScreenController factory
function was added.

As the HTML structure for the 'tic-tac-toe' game is being dynamically
added and updated via the ScreenController factory function, the
temporary markup that was in the index.html (besides the .container div)
has been removed.