Update 'resultsScreen' method to play a new game
Previously, the 'results' page was simply showing who won the game,
and if a player won, highlight the tile sequence as to how the player
won the game. But it did not allow for the players to play another game.

The 'resultsScreen' private method from the ScreenController factory
function, along with other related additions and changes to said factory
function have been made to now allow players to play a new game after
a game has been finished.

Consequently, the temporary markup that was constructed to style the
'results' page has been removed, since the changes to the
ScreenController now dynamically generate the DOM to create the same
markup once a game has finished.