Add HTML markup to create a 'start-screen'
To provide users the ability to use a custom name before playing a game,
a form was added to allow users to submit their own name. Other HTML
markup was added to construct a 'start-screen' for the game.

The ScreenController was modified to not run the 'updateScreen' method
on object creation, as it would overwrite the'start-screen' HTML markup.
In future, the plan is to run 'updateScreen' when the players submit the
form at the 'start-screen'.