Implement functional 'start-screen' page
Previously, the 'tic-tac-toe' game would immediately go straight into a
game without allowing players to change their name beforehand. The
index.html was updated in a previous commit to set up the structure to
allow players to submit custom names before playing, but the JS was not
added yet for the start-screen to be functional.

Now, the ScreenController factory function has been updated so that
players are now able to submit a custom name if desired, and the
submitted player names are now shown during the game, and even carry
through to a new game after a game has been finished.